kensho_service/requirements.txt
click==6.7
grift==0.6.0
grpcio==1.7.0
grpcio-tools==1.7.0
protobuf==3.5.0.post1
psycopg2==2.7.3.2
PyJWT==1.5.3
requests==2.13.0
schematics==1.1.1
six==1.11.0
SQLAlchemy==1.1.15
